move components to policyItemTable
