+++ src/components/Datatable/PolicyItems/BatchItemDelete.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import InfoModal from './InfoModal.svelte'
+import InfoModal from '../../InfoModal.svelte'
 import { Button, Dialog } from '@silintl/ui-components'
 import { createEventDispatcher } from 'svelte'
